API

Finding steady states

Homotopy Continuation

Analyze solutions

Access steady states

Classify steady states

Steady state plotting

Linear response

Linear response plotting

Limit-cycle methods

OrdinaryDiffEq

SteadyStateSweep

Finding steady states

HarmonicSteadyState.get_steady_states — Function

get_steady_states(
    prob::HomotopyContinuationProblem,
    method::HomotopyContinuationMethod;
    show_progress=true,
    sorting="nearest",
    classify_default=true,
    verbose=false
    )

Solves problem with the method over the ranges specified by swept_parameters, keeping fixed_parameters constant. swept_parameters accepts pairs mapping symbolic variables to arrays or ranges. fixed_parameters accepts pairs mapping symbolic variables to numbers.

Keyword arguments

show_progress: Indicate whether a progress bar should be displayed.

sorting: the method used by sort_solutions to get continuous solutions branches. The current options are "hilbert" (1D sorting along a Hilbert curve), "nearest" (nearest-neighbor sorting) and "none".

classify_default: If true, the solutions will be classified using the default classification method.

Example

solving a simple harmonic oscillator $m \ddot{x} + γ \dot{x} + ω_0^2 x = F \cos(ωt)$ to obtain the response as a function of $ω$

# having obtained a HomotopyContinuationProblem object, let's find steady states
julia> range = (ω => range(0.8, 1.2, 100) ) # 100 parameter sets to solve
julia> fixed = ParameterList(m => 1, γ => 0.01, F => 0.5, ω_0 => 1)
julia> get_steady_states(problem, range, fixed)

A steady state result for 100 parameter points

    Solution branches:   1
       of which real:    1
       of which stable:  1

    Classes: stable, physical, Hopf, binary_labels

It is also possible to perform 2-dimensional sweeps.

# The swept parameters take precedence over fixed -> use the same fixed
julia> range = (ω => range(0.8,1.2,100), F => range(0.1,1.0,10) )

# The swept parameters take precedence over fixed -> the F in fixed is now ignored
julia> get_steady_states(problem, range, fixed)

A steady state result for 1000 parameter points

    Solution branches:   1
       of which real:    1
       of which stable:  1

    Classes: stable, physical, Hopf, binary_labels

HarmonicSteadyState.SteadyStateProblem — Type

SteadyStateProblem

Abstract type for steady state problems.

HarmonicSteadyState.SteadyStateMethod — Type

SteadyStateMethod

Abstract type for steady state methods.

HarmonicSteadyState.Result — Type

mutable struct Result{D, SolType<:Number, ParType<:Number, F<:FunctionWrappers.FunctionWrapper{Array{SolType<:Number, 2}, Tuple{Array{SolType<:Number, 1}}}}

Stores the steady states of a HarmonicEquation.

Fields

solutions::Array{Array{Vector{SolType}, 1}, D} where {D, SolType<:Number}: The variable values of steady-state solutions.

swept_parameters::OrderedCollections.OrderedDict{Symbolics.Num, Vector{ParType}} where ParType<:Number: Values of all parameters for all solutions.

fixed_parameters::OrderedCollections.OrderedDict{Symbolics.Num, ParType} where ParType<:Number: The parameters fixed throughout the solutions.

problem::HarmonicSteadyState.HomotopyContinuationProblem{ParType, F} where {SolType<:Number, ParType<:Number, F<:FunctionWrappers.FunctionWrapper{Matrix{SolType}, Tuple{Vector{SolType}}}}: The HomotopyContinuationProblem used to generate this.

classes::Dict{String, Array{BitVector, D}} where D: Maps strings such as "stable", "physical" etc to arrays of values, classifying the solutions (see method classify_solutions!).

binary_labels::Array{Int64}: Create binary classification of the solutions, such that each solution point receives an identifier based on its permutation of stable branches (allows to distinguish between different phases, which may have the same number of stable solutions). It works by converting each bitstring [is_stable(solution_1), is_stable(solution_2), ...,] into unique labels.

jacobian::FunctionWrappers.FunctionWrapper{Matrix{SolType}, Tuple{Vector{SolType}}} where SolType<:Number: The Jacobian function with fixed_parameters already substituted. Accepts a vector specifying the solution. If problem.jacobian is a symbolic matrix, this holds a compiled function.

seed::UInt32: Seed used for the solver

Homotopy Continuation

HarmonicSteadyState.HomotopyContinuationProblem — Type

mutable struct HomotopyContinuationProblem{ParType<:Number, Jac<:FunctionWrappers.FunctionWrapper{Matrix{ComplexF64}, Tuple{Vector{ComplexF64}}}} <: HarmonicSteadyState.SteadyStateProblem

Holds a set of algebraic equations describing the steady state of a system.

Fields

variables::Vector{Symbolics.Num}: The harmonic variables to be solved for.

parameters::Vector{Symbolics.Num}: All symbols which are not the harmonic variables.

swept_parameters::OrderedCollections.OrderedDict{Symbolics.Num, Vector{ParType}} where ParType<:Number: The swept parameters in the homotopy.

fixed_parameters::OrderedCollections.OrderedDict{Symbolics.Num, ParType} where ParType<:Number: The fixed parameters in the homotopy.

system::HomotopyContinuation.ModelKit.System: The input object for HomotopyContinuation.jl solver methods.

jacobian::FunctionWrappers.FunctionWrapper{Matrix{ComplexF64}, Tuple{Vector{ComplexF64}}}: The Jacobian matrix (possibly symbolic or compiled function). If Matrix{Nan} and implicit function is compiled when a Result is created.

eom::QuestBase.HarmonicEquation: The HarmonicEquation object used to generate this HomotopyContinuationProblem.

Constructors

HomotopyContinuationProblem(
    eom::HarmonicEquation,
    swept::AbstractDict,
    fixed::AbstractDict;
    compile_Jacobian::Bool=true,
)

HarmonicSteadyState.HomotopyContinuationMethod — Type

HomotopyContinuationMethod

Abstract type for the steady state methods using HomotopyContinuation.jl.

HarmonicSteadyState.WarmUp — Type

WarmUp

The Warm Up method prepares a warmup system with the Total Degree method using the parameter at index perturbed by perturbation_size. The warmup system is used to perform a homotopy using all other systems in the parameter sweep. It is very efficient for systems with minimal bifurcation in the parameter sweep. The Warm Up method should in theory guarantee to find all solutions, however, if the start_parameters is not proper (to close to the real line) it could miss some solutions.

SeeHomotopyContinuation.jl for more information.

Fields

warm_up_method::Union{Polyhedral{T}, TotalDegree{T}} where T: Method used for the warmup system.

start_parameters::Vector: Start parameters.

thread::Bool: Boolean indicating if threading is enabled.

check_zero::Bool: Check if zero is a root

tracker_options::HomotopyContinuation.TrackerOptions: Options for the tracker.

endgame_options::HomotopyContinuation.EndgameOptions: Options for the endgame.

compile::Union{Bool, Symbol}: Compilation options.

seed::UInt32: Seed for random number generation.

HarmonicSteadyState.TotalDegree — Type

TotalDegree

The Total Degree homotopy method performs a homotopy $H(x, t) = γ t G(x) + (1-t) F(x)$ from the trivial polynomial system $F(x) =xᵢ^{dᵢ} +aᵢ$ with the maximal degree $dᵢ$ determined by the Bezout bound. The method guarantees to find all solutions, however, it comes with a high computational cost. See HomotopyContinuation.jl for more information.

Fields

gamma::Complex: Complex multiplying factor of the start system G(x) for the homotopy

thread::Bool: Boolean indicating if threading is enabled.

tracker_options::HomotopyContinuation.TrackerOptions: Options for the tracker.

endgame_options::HomotopyContinuation.EndgameOptions: Options for the endgame.

compile::Union{Bool, Symbol}: Compilation options.

seed::UInt32: Seed for random number generation.

HarmonicSteadyState.Polyhedral — Type

Polyhedral

The Polyhedral homotopy method constructs a homotopy based on the polyhedral structure of the polynomial system. It is more efficient than the Total Degree method for sparse systems, meaning most of the coefficients are zero. It can be especially useful if you don't need to find the zero solutions (only_non_zero = true), resulting in a speed up. See HomotopyContinuation.jl for more information.

Fields

only_non_zero::Bool: Boolean indicating if only non-zero solutions are considered.

thread::Bool: Boolean indicating if threading is enabled.

tracker_options::HomotopyContinuation.TrackerOptions: Options for the tracker.

endgame_options::HomotopyContinuation.EndgameOptions: Options for the endgame.

compile::Union{Bool, Symbol}: Compilation options.

seed::UInt32: Seed for random number generation.

Analyze solutions

Access steady states

HarmonicSteadyState.get_solutions — Function

get_solutions(
    res::Result, x::String;
    branches=1:branch_count(res), realify=false, class=["stable"], not_class=[]
    )
get_solutions(res::Result; branches=1:branch_count(res), class=["stable"], not_class=[])

Extract solution vectors from a Result object based on specified filtering criteria given by the class keywords. The first method allows extracting a specific solution component by name x. The second method returns complete solution vectors.

Keyword arguments

branches=1:branch_count(res): Range of branches to include in the output

realify=false: Whether to convert complex solutions to real form

class=["physical", "stable"]: Array of classification labels to include

not_class=[]: Array of classification labels to exclude

Returns

Filtered solution vectors matching the specified criteria

HarmonicSteadyState.get_single_solution — Function

get_single_solution(
    res::HarmonicSteadyState.Result{D, S, P, F} where F<:FunctionWrappers.FunctionWrapper{Array{S, 2}, Tuple{Array{S, 1}}};
    branch,
    index
)

Return an ordered dictionary specifying all variables and parameters of the solution in result on branch at the position index.

HarmonicSteadyState.attractors — Function

attractors(res::Result{D}; class="stable", not_class=[]) where D

Extract attractors from a Result object. Returns an array of dictionaries, where each dictionary maps branch identifier to the attractor. The attractors are filtered by their corresponding class.

Keyword arguments

Class selection done by passing String or Vector{String} as kwarg:

class::String       :   only count solutions in this class ("all" --> plot everything)
not_class::String   :   do not count solutions in this class

Returns

Array{Dict,D}: Vector of dictionaries mapping branch indices to points satisfying the stability criteria at each parameter value

HarmonicSteadyState.transform_solutions — Function

transform_solutions(
    res::HarmonicSteadyState.Result{D, S, ParType, F} where {ParType<:Number, F<:FunctionWrappers.FunctionWrapper{Array{S, 2}, Tuple{Array{S, 1}}}},
    func;
    branches,
    realify
) -> Vector

Takes a Result object and a string f representing a Symbolics.jl expression. Returns an array with the values of f evaluated for the respective solutions. Additional substitution rules can be specified in rules in the format ("a" => val) or (a => val)

Classify steady states

HarmonicSteadyState.classify_solutions! — Function

classify_solutions!(
    res::HarmonicSteadyState.Result,
    func::Union{Function, String},
    name::String;
    physical
) -> Any

Creates a solution class in res using the function func (parsed into Symbolics.jl input). The new class is labeled with name and stored under res.classes[name]. By default, only physical (real) solutions are classified, and false is returned for the rest. To also classify complex solutions, set physical=false.

Example

# solve a previously-defined problem
res = get_steady_states(problem, swept_parameters, fixed_parameters)

# classify, store in result.classes["large_amplitude"]
classify_solutions!(res, "sqrt(u1^2 + v1^2) > 1.0" , "large_amplitude")

HarmonicSteadyState.get_class — Function

get_class(
    res::HarmonicSteadyState.Result,
    branch::Int64,
    class::String
) -> Any

Returns an array of booleans classifying branch in the solutions in res according to class.

get_class(
    soln::HarmonicSteadyState.Result,
    class::String
) -> Vector

Returns an array of booleans classifying each branch in the solutions in res according to class.

HarmonicSteadyState.filter_result! — Function

filter_result!(
    res::HarmonicSteadyState.Result,
    class::String
)

Removes all solution branches from res where NONE of the solution falls into class. Typically used to filter out unphysical solutions to prevent huge file sizes.

HarmonicSteadyState.phase_diagram — Function

phase_diagram(res::Result{D}; class="physical", not_class=[]) where {D}

Calculate the phase diagram from a Result object by summing over the number of states at each swept parameters.

Keyword arguments

Class selection done by passing String or Vector{String} as kwarg:

class::String       :   only count solutions in this class ("all" --> plot everything)
not_class::String   :   do not count solutions in this class

Returns

Array{Int64,D}: Sum of states after applying the specified class masks

Steady state plotting

To use these plotting functions, you need to have the Plots package installed in the same enverionment and loaded.

RecipesBase.plot — Method

plot(
    res::HarmonicSteadyState.Result{D, S, P, F} where F<:FunctionWrappers.FunctionWrapper{Array{S, 2}, Tuple{Array{S, 1}}},
    varargs...;
    cut,
    kwargs...
) -> Plots.Plot

Plot a Result object.

Class selection done by passing String or Vector{String} as kwarg:

class       :   only plot solutions in this class(es) ("all" --> plot everything)
not_class   :   do not plot solutions in this class(es)

Other kwargs are passed onto Plots.gr().

See also plot!

The x,y,z arguments are Strings compatible with Symbolics.jl, e.g., y=2*sqrt(u1^2+v1^2) plots the amplitude of the first quadratures multiplied by 2.

1D plots

plot(res::Result; x::String, y::String, class="default", not_class=[], kwargs...)
plot(res::Result, y::String; kwargs...) # take x automatically from Result

Default behaviour is to plot stable solutions as full lines, unstable as dashed.

If a sweep in two parameters were done, i.e., dimension(res)==2, a one dimensional cut can be plotted by using the keyword cut were it takes a Pair{Num, Float} type entry. For example, plot(res, y="sqrt(u1^2+v1^2), cut=(λ => 0.2)) plots a cut at λ = 0.2.

2D plots

plot(res::Result; z::String, branch::Int64, class="physical", not_class=[], kwargs...)

To make the 2d plot less chaotic it is required to specify the specific branch to plot, labeled by a Int64.

The x and y axes are taken automatically from res

RecipesBase.plot! — Method

plot!(
    res::HarmonicSteadyState.Result,
    varargs...;
    kwargs...
) -> Plots.Plot

Similar to plot but adds a plot onto an existing plot.

HarmonicSteadyState.plot_phase_diagram — Function

plot_phase_diagram(
    res::HarmonicSteadyState.Result{D, SolType} where SolType<:Number;
    kwargs...
) -> Plots.Plot

Plot the number of solutions in a Result object as a function of the parameters. Works with 1D and 2D datasets.

Class selection done by passing String or Vector{String} as kwarg:

class::String       :   only count solutions in this class ("all" --> plot everything)
not_class::String   :   do not count solutions in this class

Other kwargs are passed onto Plots.gr()

HarmonicSteadyState.plot_spaghetti — Function

plot_spaghetti(
    res::HarmonicSteadyState.Result{D, SolType} where SolType<:Number;
    x,
    y,
    z,
    class,
    not_class,
    add,
    kwargs...
)

Plot a three dimension line plot of a Result object as a function of the parameters. Works with 1D and 2D datasets.

Class selection done by passing String or Vector{String} as kwarg:

class::String       :   only count solutions in this class ("all" --> plot everything)
not_class::String   :   do not count solutions in this class

Other kwargs are passed onto Plots.gr()

Linear response

HarmonicSteadyState.LinearResponse.eigenvalues — Method

eigenvalues(res::Result, branch; class=["physical"])

Calculate the eigenvalues of the Jacobian matrix of the harmonic equations of a branch for a one dimensional sweep in the Result struct.

Arguments

res::Result: Result object containing solutions and jacobian information

branch: Index of the solution branch to analyze

class=["physical"]: Filter for solution classes to include, defaults to physical solutions

Returns

Vector of filtered eigenvalues along the solution branch

Notes

Currently only supports 1-dimensional parameter sweeps (D=1)

Will throw an error if branch contains NaN values

Eigenvalues are filtered based on the specified solution classes

HarmonicSteadyState.LinearResponse.eigenvectors — Method

eigenvectors(res::Result, branch; class=["physical"])

get_Jacobiannch to analyze

class=["physical"]: Filter for solution classes to include, defaults to physical solutions

Returns

Vector of filtered eigenvectors along the solution branch

Notes

Currently only supports 1-dimensional parameter sweeps (D=1)

Will throw an error if branch contains NaN values

Eigenvectors are filtered based on the specified solution classes

HarmonicSteadyState.LinearResponse.get_jacobian_response — Method

get_jacobian_response(
    res::HarmonicSteadyState.Result{D, S, P, F} where F<:FunctionWrappers.FunctionWrapper{Array{S, 2}, Tuple{Array{S, 1}}},
    nat_var::Symbolics.Num,
    Ω_range,
    branch::Int64;
    show_progress
) -> Matrix

Calculate the Jacobian response spectrum for a given system. Computes the magnitude of the Jacobian response for stable solutions across specified frequency ranges.

Arguments

res::Result: Result object containing the system's solutions

nat_var::Num: Natural variable to evaluate in the response

Ω_range: Range of frequencies to evaluate

branch::Int or followed_branches::Vector{Int}: Branch number(s) to analyze

show_progress=true: Whether to show a progress bar

force=false: Force recalculation of spectrum even if already exists

Returns

Array{P,2}: Complex response matrix where rows correspond to frequencies and columns to solutions

HarmonicSteadyState.LinearResponse.get_linear_response — Method

get_linear_response(
    res::HarmonicSteadyState.Result{D, S, P, F} where F<:FunctionWrappers.FunctionWrapper{Array{S, 2}, Tuple{Array{S, 1}}},
    nat_var::Symbolics.Num,
    Ω_range,
    branch::Int64;
    show_progress
)

Calculate the linear response of the system for a given branch. Evaluates the linear response by solving the linear response ODE for each stable solution and input frequency in the given range.

Arguments

res: Result object containing the system's solutions

nat_var::Num: Natural variable to evaluate in the response

Ω_range: Range of frequencies to evaluate

branch::Int: Branch number to analyze

show_progress=true: Whether to show a progress bar

Returns

Array{P,2}: Response matrix where rows correspond to frequencies and columns to stable solutions

HarmonicSteadyState.LinearResponse.get_rotframe_jacobian_response — Method

get_rotframe_jacobian_response(
    res::HarmonicSteadyState.Result{D, S, P, F} where F<:FunctionWrappers.FunctionWrapper{Array{S, 2}, Tuple{Array{S, 1}}},
    Ω_range,
    branch::Int64;
    show_progress,
    damping_mod
)

Calculate the rotating frame Jacobian response for a given branch. Computes the rotating frame Jacobian response by evaluating eigenvalues of the numerical Jacobian and calculating the response magnitude for each frequency in the range.

Arguments

res::Result: Result object containing the system's solutions

Ω_range: Range of frequencies to evaluate

branch::Int: Branch number to analyze

show_progress=true: Whether to show a progress bar

damping_mod: Damping modification parameter

Returns

Array{P,2}: Response matrix in the rotating frame

Linear response plotting

HarmonicSteadyState.plot_eigenvalues — Function

plot_eigenvalues(
    res::HarmonicSteadyState.Result{D, S, P, F} where F<:FunctionWrappers.FunctionWrapper{Array{S, 2}, Tuple{Array{S, 1}}},
    branch::Int64;
    class,
    type,
    projection,
    cscheme,
    kwargs...
) -> Any

Visualize the eigenvalues of the Jacobian in the rotating frame for branch identifier in the Result res.

Keyword arguments

class: Array of solution classes to include (default: ["physical"])

type: Which part of eigenvalues to plot (:real or :imag, default: :imag)

projection: Function mapping eigenvectors to colors (default: v->1)

cscheme: Color scheme for plotting (:default or custom scheme)

kwargs...: Additional arguments passed to Plots.scatter

Returns

A scatter plot of eigenvalues colored by the projection of their eigenvectors.

Example

# Plot imaginary parts of eigenvalues
plot_eigenvalues(result, branch=1)

# Plot real parts with custom coloring based on the norm of eigenvectors of the first harmonic
plot_eigenvalues(result, branch=1, type=:real, projection=v->sqrt(v[1]^2+v[2]^2))

HarmonicSteadyState.plot_linear_response — Function

plot_linear_response(
    res::HarmonicSteadyState.Result{D, SolType} where SolType<:Number,
    nat_var::Symbolics.Num,
    branch::Int64;
    Ω_range,
    order,
    logscale,
    show_progress,
    kwargs...
)

Plot the linear response to white noise of the variable nat_var for Result res on branch identifier.

Keyword arguments

Ω_range: Range of frequency of the noise probe

order: Order of slow-time derivatives to keep (default: 1)

logscale: Whether to plot response in log scale (default: false)

show_progress: Show progress bar during computation (default: true)

kwargs...: Additional arguments passed to Plots.heatmap

Returns

A Plots.jl heatmap showing the linear response magnitude across parameter and frequency space.

plot_linear_response(
    res::HarmonicSteadyState.Result,
    nat_var::Symbolics.Num,
    followed_branches::Vector{Int64};
    Ω_range,
    logscale,
    show_progress,
    switch_axis,
    force,
    kwargs...
)

Plot the linear response to white noise of the variable nat_var for Result res on the followed_branches identifiers with the size of Ω_range.

Keyword arguments

Ω_range: Range of frequency of the noise probe

order: Order of slow-time derivatives to keep (default: 1)

logscale: Whether to plot response in log scale (default: false)

show_progress: Show progress bar during computation (default: true)

kwargs...: Additional arguments passed to Plots.heatmap

Returns

A Plots.jl heatmap showing the linear response magnitude across parameter and frequency space.

HarmonicSteadyState.plot_rotframe_jacobian_response — Function

plot_rotframe_jacobian_response(
    res::HarmonicSteadyState.Result{D, S, P, F} where F<:FunctionWrappers.FunctionWrapper{Array{S, 2}, Tuple{Array{S, 1}}},
    branch::Int64;
    Ω_range,
    logscale,
    damping_mod,
    show_progress,
    kwargs...
)

Plot the linear response to white noise in the rotating frame defined the harmonic ansatz for Result res on branch identifier.

Keyword arguments

Ω_range: Range of frequencies to analyze

logscale: Whether to plot response in log scale (default: true)

damping_mod: Multiplier for the real part of Jacobian eigenvalues (default: 1.0)

show_progress: Show progress bar during computation (default: true)

kwargs...: Additional arguments passed to Plots.heatmap

Returns

A Plots.jl heatmap showing the response magnitude in the rotating frame.

Notes

Setting damping_mod < 1 can help distinguish between peaks with similar frequencies

Solutions not belonging to the physical class are ignored

Limit-cycle methods

HarmonicSteadyState.LimitCycles.get_limit_cycles — Function

get_limit_cycles(
    eom::HarmonicEquation, method::SteadyStateMethod, swept, fixed, ω_lc; kwargs...)

Variant of get_steady_states for a limit cycle problem characterised by a Hopf frequency (usually called ω_lc)

Solutions with ω_lc = 0 are labelled unphysical since this contradicts the assumption of distinct harmonic variables corresponding to distinct harmonics.

HarmonicSteadyState.LimitCycles.add_pairs! — Function

add_pairs!(eom::DifferentialEquation; ω_lc::Num, n=1)

Add a limit cycle harmonic ω_lc to the system Equivalent to adding n pairs of harmonics ω +- ω_lc for each existing ω.

HarmonicSteadyState.LimitCycles.get_cycle_variables — Function

get_cycle_variables(
    eom::QuestBase.HarmonicEquation,
    ω_lc::Symbolics.Num
) -> Vector{QuestBase.HarmonicVariable}

Return the harmonic variables which participate in the limit cycle labelled by ω_lc.

OrdinaryDiffEq

HarmonicSteadyState.AdiabaticSweep — Type

Represents a sweep of one or more parameters of a HarmonicEquation. During a sweep, the selected parameters vary linearly over some timespan and are constant elsewhere.

Sweeps of different variables can be combined using +.

Fields

functions::Dict{Symbolics.Num, Function}: Maps each swept parameter to a function.

Examples

# create a sweep of parameter a from 0 to 1 over time 0 -> 100
julia> @variables a,b;
julia> sweep = AdiabaticSweep(a => [0., 1.], (0, 100));
julia> sweep[a](50)
0.5
julia> sweep[a](200)
1.0

# do the same, varying two parameters simultaneously
julia> sweep = AdiabaticSweep([a => [0.,1.], b => [0., 1.]], (0,100))

Successive sweeps can be combined,

sweep1 = AdiabaticSweep(ω => [0.95, 1.0], (0, 2e4))
sweep2 = AdiabaticSweep(λ => [0.05, 0.01], (2e4, 4e4))
sweep = sweep1 + sweep2

multiple parameters can be swept simultaneously,

sweep = AdiabaticSweep([ω => [0.95;1.0], λ => [5e-2;1e-2]], (0, 2e4))

and custom sweep functions may be used.

ωfunc(t) = cos(t)
sweep = AdiabaticSweep(ω => ωfunc)

HarmonicSteadyState.follow_branch — Function

follow_branch(
    starting_branch::Int64,
    res::HarmonicSteadyState.Result;
    y,
    sweep,
    tf,
    ϵ
) -> Tuple{Any, Any}

Return the indexes and values following stable branches along a 1D sweep. When a no stable solutions are found (e.g. in a bifurcation), the next stable solution is calculated by time evolving the previous solution (quench).

Keyword arguments

y: Dependent variable expression (parsed into Symbolics.jl) to evaluate the followed solution branches on .

sweep: Direction for the sweeping of solutions. A right (left) sweep proceeds from the first (last) solution, ordered as the sweeping parameter.

tf: time to reach steady

ϵ: small random perturbation applied to quenched solution, in a bifurcation in order to favour convergence in cases where multiple solutions are identically accessible (e.g. symmetry breaking into two equal amplitude states)

HarmonicSteadyState.plot_1D_solutions_branch — Function

plot_1D_solutions_branch(
    starting_branch::Int64,
    res::HarmonicSteadyState.Result;
    x,
    y,
    sweep,
    tf,
    ϵ,
    class,
    not_class,
    kwargs...
)

Plot a bifurcation diagram from a continuation sweep starting from starting_branch using the Result struct res. Time integration is used to determine what follow up branch in the continuation is.

Keyword arguments

x::String: Expression for the x-axis variable

y::String: Expression for the y-axis variable

sweep::String="right": Direction to follow the branch ("right" or "left")

tf::Real=10000: Final time for time integration

ϵ::Real=1e-4: Tolerance for branch following

kwargs...: Additional plotting arguments passed to Plots.jl

Class selection done by passing String or Vector{String} as kwarg:

class::String : only count solutions in this class ("all" –> plot everything) not_class::String : do not count solutions in this class

Returns

A Plots.jl plot object containing the bifurcation diagram with the followed branch

Description

This function creates a bifurcation diagram using follow_branch. The followed branch is plotted as a dashed gray line.

HarmonicSteadyState.follow_branch — Method

follow_branch(
    starting_branch::Int64,
    res::HarmonicSteadyState.Result;
    y,
    sweep,
    tf,
    ϵ
) -> Tuple{Any, Any}

Return the indexes and values following stable branches along a 1D sweep. When a no stable solutions are found (e.g. in a bifurcation), the next stable solution is calculated by time evolving the previous solution (quench).

Keyword arguments

y: Dependent variable expression (parsed into Symbolics.jl) to evaluate the followed solution branches on .

sweep: Direction for the sweeping of solutions. A right (left) sweep proceeds from the first (last) solution, ordered as the sweeping parameter.

tf: time to reach steady

ϵ: small random perturbation applied to quenched solution, in a bifurcation in order to favour convergence in cases where multiple solutions are identically accessible (e.g. symmetry breaking into two equal amplitude states)

SteadyStateSweep

HarmonicSteadyState.steady_state_sweep — Function

steady_state_sweep(prob::SteadyStateProblem, alg::DynamicSS; varied::Pair, kwargs...)

Sweeps through a range of parameter values using a dynamic steady state solver DynamicSS of the SteadyStateDiffEq.jl package. Given a steady state problem and a parameter to vary, computes the steady state solution for each value in the sweep range. The solutions are returned as a vector where each element corresponds to the steady state found at that parameter value.

steady_state_sweep(prob_np::NonlinearProblem, prob_ss::SteadyStateProblem,
                  alg_np, alg_ss::DynamicSS; varied::Pair, kwargs...)

Performs a parameter sweep by combining nonlinear root alg_np and steady state solvers alg_ss. For each parameter value, it first attempts a direct nonlinear root solver and checks its stability. If the solution is unstable or not found, it switches to a dynamic steady state solver. This hybrid approach is much faster then only using a steady state solver.